+++ startup/launcher/src/commonMain/kotlin/StartLauncherPlugin.kt
@@ -4,6 +4,8 @@ import dev.inmo.kslog.common.i
 import dev.inmo.kslog.common.taggedLogger
 import dev.inmo.kslog.common.w
 import dev.inmo.micro_utils.coroutines.runCatchingSafely
+import dev.inmo.micro_utils.startup.launcher.StartLauncherPlugin.setupDI
+import dev.inmo.micro_utils.startup.launcher.StartLauncherPlugin.startPlugin
 import dev.inmo.micro_utils.startup.plugin.StartPlugin
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -12,7 +14,10 @@ import kotlinx.coroutines.launch
 import kotlinx.serialization.SerialFormat
 import kotlinx.serialization.StringFormat
 import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.jsonObject
 import org.koin.core.Koin
+import org.koin.core.KoinApplication
+import org.koin.core.context.startKoin
 import org.koin.core.module.Module
 import org.koin.dsl.binds
 import org.koin.dsl.module
@@ -23,26 +28,20 @@ import org.koin.dsl.module
 object StartLauncherPlugin : StartPlugin {
     internal val logger = taggedLogger(this)
 
-    /**
-     * Will deserialize [Config] from [config], register it in receiver [Module] (as well as [CoroutineScope] and
-     * [kotlinx.serialization.json.Json])
-     *
-     * Besides, in this method will be called [StartPlugin.setupDI] on each plugin from [Config.plugins]. In case when
-     * some plugin will not be loaded correctly it will be reported throw the [logger]
-     */
-    override fun Module.setupDI(config: JsonObject) {
-        val pluginsConfig = defaultJson.decodeFromJsonElement(Config.serializer(), config)
+    fun Module.setupDI(config: Config, rawJsonObject: JsonObject? = null) {
+        val rawJsonObject = rawJsonObject ?: defaultJson.encodeToJsonElement(Config.serializer(), config).jsonObject
 
-        single { pluginsConfig }
+        single { rawJsonObject }
+        single { config }
         single { CoroutineScope(Dispatchers.Default) }
         single { defaultJson } binds arrayOf(StringFormat::class, SerialFormat::class)
 
         includes(
-            pluginsConfig.plugins.mapNotNull {
+            config.plugins.mapNotNull {
                 runCatching {
                     module {
                         with(it) {
-                            setupDI(config)
+                            setupDI(rawJsonObject)
                         }
                     }
                 }.onFailure { e ->
@@ -53,10 +52,27 @@ object StartLauncherPlugin : StartPlugin {
     }
 
     /**
+     * Will deserialize [Config] from [config], register it in receiver [Module] (as well as [CoroutineScope] and
+     * [kotlinx.serialization.json.Json])
+     *
+     * Besides, in this method will be called [StartPlugin.setupDI] on each plugin from [Config.plugins]. In case when
+     * some plugin will not be loaded correctly it will be reported throw the [logger]
+     */
+    override fun Module.setupDI(config: JsonObject) {
+        logger.i("Koin for current module has started setup")
+        setupDI(
+            defaultJson.decodeFromJsonElement(Config.serializer(), config),
+            config
+        )
+        logger.i("Koin for current module has been setup")
+    }
+
+    /**
      * Takes [CoroutineScope] and [Config] from the [koin], and call starting of each plugin from [Config.plugins]
      * ASYNCHRONOUSLY. Just like in [setupDI], in case of fail in some plugin it will be reported using [logger]
      */
     override suspend fun startPlugin(koin: Koin) {
+        logger.i("Start starting of subplugins")
         val scope = koin.get<CoroutineScope>()
         koin.get<Config>().plugins.map { plugin ->
             scope.launch {
@@ -66,11 +82,60 @@ object StartLauncherPlugin : StartPlugin {
                         startPlugin(koin)
                     }
                 }.onFailure { e ->
-                    logger.w("Unable to load bot part of $plugin", e)
+                    logger.w("Unable to start plugin $plugin", e)
                 }.onSuccess {
                     logger.i("Complete loading of $plugin")
                 }
             }
         }.joinAll()
+        logger.i("Complete subplugins start")
+    }
+
+    /**
+     * Will create [KoinApplication], init, load modules using [StartLauncherPlugin] and start plugins using the same base
+     * plugin
+     *
+     * @param rawConfig It is expected that this [JsonObject] will contain serialized [Config] ([StartLauncherPlugin] will
+     * deserialize it in its [StartLauncherPlugin.setupDI]
+     */
+    suspend fun start(rawConfig: JsonObject) {
+
+        logger.i("Start initialization")
+        val koinApp = KoinApplication.init()
+        koinApp.modules(
+            module {
+                setupDI(rawConfig)
+            }
+        )
+        logger.i("Modules loaded")
+        startKoin(koinApp)
+        logger.i("Koin started")
+        startPlugin(koinApp.koin)
+        logger.i("App has been setup")
+
+    }
+
+    /**
+     * Will create [KoinApplication], init, load modules using [StartLauncherPlugin] and start plugins using the same base
+     * plugin
+     *
+     * @param config In difference with other [start] method here config is already deserialized and this config will
+     * be converted to [JsonObject] as raw config. That means that all plugins from [config] will receive
+     * serialized version of [config] in [StartPlugin.setupDI] method
+     */
+    suspend fun start(config: Config) {
+
+        logger.i("Start initialization")
+        val koinApp = KoinApplication.init()
+        logger.i("Koin app created")
+        koinApp.modules(
+            module {
+                setupDI(config)
+            }
+        )
+        startKoin(koinApp)
+        logger.i("Koin started")
+        startPlugin(koinApp.koin)
+
     }
 }