+++ startup/launcher/src/commonTest/kotlin/StartupLaunchingTests.kt
@@ -1,7 +1,7 @@
 import dev.inmo.micro_utils.startup.launcher.Config
 import dev.inmo.micro_utils.startup.launcher.HelloWorldPlugin
+import dev.inmo.micro_utils.startup.launcher.StartLauncherPlugin
 import dev.inmo.micro_utils.startup.launcher.defaultJson
-import dev.inmo.micro_utils.startup.launcher.start
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.json.jsonObject
@@ -23,7 +23,7 @@ class StartupLaunchingTests {
 
         runTest {
             val  job = launch {
-                start(emptyJson)
+                StartLauncherPlugin.start(emptyJson)
             }
             job.join()
         }
@@ -37,7 +37,7 @@ class StartupLaunchingTests {
 
         runTest {
             val  job = launch {
-                start(emptyJson)
+                StartLauncherPlugin.start(emptyJson)
             }
             job.join()
         }